big/backend_bigdata/pyproject.toml
[project]
name = "backend_bigdata"
version = "0.1.0"
description = "Backend service for maintenance prediction"
authors = []
readme = "README.md"
requires-python = "^3.12"
dependencies = [
    "fastapi",
    "uvicorn",
    "sqlalchemy",
    "psycopg2-binary",
    "alembic",
    "python-dotenv",
    "prometheus-client",
    "redpanda",
]
packages = [{ include = "app" }]